alura_cursos/flask_curso/requirements.txt
click==7.1.2
Flask==0.12.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
click==7.1.2
Flask==0.12.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
Werkzeug==1.0.1
